consolidate client fixture and api tests
